py-DDGR1.0/requirement.txt
absl-py==0.15.0
astunparse==1.6.3
blessings==1.7
blobfile==2.0.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.7
clang==5.0
cycler==0.10.0
filelock==3.9.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.41.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.1
jsonpatch==1.32
jsonpointer==2.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.9.2
Markdown==3.3.4
matplotlib==3.4.3
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
olefile==0.46
opencv-python==4.7.0.68
opt-einsum==3.3.0
pip==21.2.2
protobuf==3.19.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.16.0
pyparsing==2.4.7
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
setuptools==58.0.4
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
torchattacks==3.2.5
torchfile==0.1.0
torchnet==0.0.4
tornado==6.1
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
visdom==0.1.8.9
websocket-client==1.2.1
Werkzeug==2.0.2
wheel==0.37.0
wrapt==1.12.1
zipp==3.6.0
